Add files via upload
Diwali Sales Analysis Using Python

Data Analysis or sometimes referred to as exploratory data analysis (EDA) is one of the core components of data science. It is also the part on the majority of the time which makes it extremely important in the field of data science. This repository demonstrates Exploratory Data Analysis methods and techniques using Python. The purpose of the used Diwali Sales dataset has been taken from Kaggle since it is one of the ideal data set for performing EDA and taking a step towards the most amazing and interesting field of data science. Good luck with your EDA on the used Diwali Sales dataset.

Project Description: -

•	Performed Data Cleaning and Data Manipulation.
•	Performed Exploratory Data Analysis (EDA) using Pandas, NumPy, Matplotlib, Seaborn Libraries.
•	Improved Customer experience by identifying potential customers across different states, occupation, gender and age groups.
•	Improved sales by identifying most selling product categories and products, which can help to plan inventory and hence meet the demands.

Conclusion: -
•	Married women age group 26-35 yrs from UP,
•	Maharashtra and Karnataka working in IT,
•	Healthcare and Aviation are more likely to buy products from Food,
•	Clothing and Electronics category
•	Targeted marketing efforts can be tailored towards these customer segments to improve sales.